requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
Django==3.0.5
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-js-asset==1.2.2
importlib-metadata==1.6.0
kombu==4.6.8
Pillow==7.1.2
python-dotenv==0.14.0
pytz==2019.3
redis==3.4.1
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
